+++ src/LevelSearchLayer.cpp
@@ -79,7 +79,7 @@ class $modify(LevelSearchLayer) {
         
         sunrise->setZOrder(-3);
         sunrise->setAnchorPoint({0, 0});
-        sunrise->setScaleX(11 * relativescale);
+        sunrise->setScaleX(15 * relativescale);
 		sunrise->setScaleY(1.6 * relativescale);
         sunrise->setPosition({0, 0});
 		sunrise->setID("sunrise"_spr);
@@ -127,7 +127,7 @@ class $modify(LevelSearchLayer) {
         
         sunset->setZOrder(-3);
         sunset->setAnchorPoint({0, 0});
-        sunset->setScaleX(11 * relativescale);
+        sunset->setScaleX(15 * relativescale);
 		sunset->setScaleY(1.6 * relativescale);
         sunset->setPosition({0, 0});
 		sunset->setID("sunset"_spr);