+++ src/ShareCommentLayer.cpp
@@ -1,0 +1,20 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/ShareCommentLayer.hpp>
+#include "Utils.hpp"
+
+using namespace geode::prelude;
+
+class $modify(ShareCommentLayer) {
+
+	bool init(gd::string title, int charLimit, CommentType type, int ID, gd::string desc) {
+		if (!ShareCommentLayer::init(title, charLimit, type, ID, desc))
+		return false;
+
+	auto DayAndNightSystemOverlay = DayAndNightSystemOverlay::create();
+	DayAndNightSystemOverlay->setZOrder(106);
+	DayAndNightSystemOverlay->setID("ScreenOverlay"_spr);
+	this->addChild(DayAndNightSystemOverlay);
+
+		return true;
+	}
+};